fix tests, added booking and rating

+++ src/main/java/edu/uoc/abarrena/trips/infrastructure/rest/dto/response/CruiseDto.java
@@ -12,9 +12,9 @@ public class CruiseDto {
     private Long id;
     private String name;
     private String description;
-    private int capacity;
-    private float avgShipRating;
-    private float avgRoomRating;
-    private float avgCrewRating;
-    private float avgFoodRating;
+    private Integer capacity;
+    private Double avgShipRating;
+    private Double avgRoomRating;
+    private Double avgCrewRating;
+    private Double avgFoodRating;
 }
